% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contab.h.R
\name{contab}
\alias{contab}
\title{Independent Samples}
\usage{
contab(
  data,
  rows,
  cols,
  counts = NULL,
  correction = "ob",
  obs = TRUE,
  exp = FALSE,
  sr = FALSE,
  ss = FALSE,
  pcRow = FALSE,
  pcCol = FALSE,
  pcTot = FALSE,
  varA = FALSE,
  text = TRUE,
  barplot = FALSE,
  yaxis = "ycounts",
  yaxisPc = "total_pc",
  xaxis = "xrows",
  bartype = "dodge",
  formula
)
}
\arguments{
\item{data}{the data as a data frame}

\item{rows}{the variable to use as the rows in the contingency table}

\item{cols}{the variable to use as the columns in the contingency table}

\item{counts}{the variable for counts, required}

\item{correction}{correction for parameters, none, Occam's bonus (default)
or AIC}

\item{obs}{\code{TRUE} or \code{FALSE} (default), provide the observed
counts}

\item{exp}{\code{TRUE} or \code{FALSE} (default), provide the expected
counts}

\item{sr}{\code{TRUE} or \code{FALSE} (default), provide the standardized
residuals}

\item{ss}{\code{TRUE} or \code{FALSE} (default), provide the signed S
values}

\item{pcRow}{\code{TRUE} or \code{FALSE} (default), provide row percentages}

\item{pcCol}{\code{TRUE} or \code{FALSE} (default), provide column
percentages}

\item{pcTot}{\code{TRUE} or \code{FALSE} (default), provide total
percentages}

\item{varA}{\code{TRUE} or \code{FALSE} (default), perform variance
analysis against the null model}

\item{text}{\code{TRUE} (default) or \code{FALSE}, how to report the
results}

\item{barplot}{\code{TRUE} or \code{FALSE} (default), show barplots}

\item{yaxis}{ycounts (default) or ypc. Use respectively \code{counts} or
\code{percentages} for the bar plot y-axis}

\item{yaxisPc}{total_pc (default), column_pc, or row_pc. Use respectively
percentages \code{of total}, \code{within columns}, or \code{within rows}
for the bar plot y-axis.}

\item{xaxis}{rows (default), or columns in bar plot X axis}

\item{bartype}{stack or side by side (default), barplot type}

\item{formula}{(optional) the formula to use, see the examples}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$freqs} \tab \tab \tab \tab \tab a proportions table \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$cttma} \tab \tab \tab \tab \tab a table \cr
  \code{results$ctt3} \tab \tab \tab \tab \tab a table \cr
  \code{results$barplot} \tab \tab \tab \tab \tab an image \cr
  \code{results$tabText} \tab \tab \tab \tab \tab a html \cr
  \code{results$SupportTab} \tab \tab \tab \tab \tab a table \cr
  \code{results$MoretabText} \tab \tab \tab \tab \tab a html \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$freqs$asDF}

\code{as.data.frame(results$freqs)}
}
\description{
Two-way analysis of a contingency table where main effects and interaction 
can be assessed.
}
\examples{
dat <- data.frame(
  Infested = c('Yes', 'Yes', 'Yes', 'Yes', 'Yes', 'No', 'No', 'No', 'No', 'No'),
  Age = c(1,2,3,4,5),
  Eggs = c(14, 16, 14, 7, 6, 87, 33, 66, 34, 11),
  check.names=FALSE)

jeva::contab(formula = Eggs ~ Infested:Age, data = dat, text = FALSE)
#
# INDEPENDENT SAMPLES
#
# Contingency Table
# ----------------------------------------------------
#  Infested    1      2     3     4     5     Total
# ----------------------------------------------------
#   No           87    33    66    34    11      231
#   Yes          14    16    14     7     6       57
#   Total       101    49    80    41    17      288
# ----------------------------------------------------
#
#
# Support: Marginal main effects and interaction analyses, against the Null model
# ----------------------------------------------------------------------------------------------
#   Component           Expected value    S            Param    G             df    p
# ----------------------------------------------------------------------------------------------
#   Infested                 144.00000    55.846183     1, 2    112.692366     1    < .0000001
#   Age                       57.60000    38.395080     1, 5     80.790159     4    < .0000001
#   Infested  тип  Age                       2.806653     1, 5      9.613306     4     0.0474704
#   Total                     28.80000    97.047916    1, 10    203.095832     9    < .0000001
# ----------------------------------------------------------------------------------------------
#   Note. S uses Occam's Bonus correction for parameters (Param).
#

}
