% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2sttest.h.R
\name{l2sttest}
\alias{l2sttest}
\title{Independent Samples T-Test}
\usage{
l2sttest(
  data,
  dep,
  group,
  welch = TRUE,
  nul = 0,
  alt = 0,
  lint = 2,
  correction = "ob",
  pll = FALSE,
  plotype = "lplot",
  supplot = -10,
  dtab = FALSE,
  plt = FALSE,
  text = TRUE
)
}
\arguments{
\item{data}{the data as a data frame}

\item{dep}{a (non-empty) numeric vector of data values}

\item{group}{an integer vector the same length as data, coding for 2 groups}

\item{welch}{\code{TRUE} (default) or \code{FALSE}, perform Welch's t-test}

\item{nul}{difference value for the null hypothesis, default = 0}

\item{alt}{difference value for an alternative hypothesis, default = 0}

\item{lint}{likelihood interval given as support value, e.g. 2 or 3,
default = 2}

\item{correction}{correction for parameters, none, Occam's bonus (default)
or AIC}

\item{pll}{\code{TRUE} or \code{FALSE} (default), give the likelihood or
log likelihood function showing  null hypothesis (black line), alternative
hypothesis (blue line), mean (dashed line),  and specified support interval
(horizontal red line)}

\item{plotype}{choose type of plot, likelihood function (default), support
function}

\item{supplot}{To set the minimum likelihood display value in plot, in log
units (default = -10)  affects the x-axis range}

\item{dtab}{\code{TRUE} or \code{FALSE} (default), provide descriptive
statistics}

\item{plt}{\code{TRUE} or \code{FALSE} (default), give a plot of means with
specified support intervals}

\item{text}{\code{TRUE} (default) or \code{FALSE}, how to report the
results}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$l2sttest} \tab \tab \tab \tab \tab a table \cr
  \code{results$l2sttest2} \tab \tab \tab \tab \tab a table \cr
  \code{results$l2sttestd} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plotc} \tab \tab \tab \tab \tab an image \cr
  \code{results$tabText} \tab \tab \tab \tab \tab a html \cr
  \code{results$SupportTab} \tab \tab \tab \tab \tab a table \cr
  \code{results$MoretabText} \tab \tab \tab \tab \tab a html \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$l2sttest$asDF}

\code{as.data.frame(results$l2sttest)}
}
\description{
Independent samples analysis where null and alternative hypotheses canbe 
specified. Likelihood interval support level can be specified and plotted 
with the means. A likelihood function for the difference in means can be 
plotted.
}
\examples{
data('ToothGrowth')
jeva::l2sttest(data = ToothGrowth, dep = len, group = supp, text = FALSE)

}
