% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2waov.h.R
\name{l2waov}
\alias{l2waov}
\title{Two-Way Factorial ANOVA}
\usage{
l2waov(
  data,
  dep,
  factor1 = NULL,
  factor2 = NULL,
  ct1 = FALSE,
  Ct1Values = "1,-1,...",
  ct2 = FALSE,
  Ct2Values = "1,-1,...",
  correction = "ob",
  plt = FALSE,
  ME = FALSE,
  text = TRUE,
  formula
)
}
\arguments{
\item{data}{the data as a data frame}

\item{dep}{the dependent variable from \code{data}, variable must be
numeric (not necessary when providing a formula, see examples)}

\item{factor1}{the 1st grouping or independent variable}

\item{factor2}{the 2nd grouping or independent variable}

\item{ct1}{\code{TRUE} or \code{FALSE} (default), for contrast 1}

\item{Ct1Values}{a comma-separated list specifying the contrast}

\item{ct2}{\code{TRUE} or \code{FALSE} (default), for contrast 2}

\item{Ct2Values}{a comma-separated list specifying the contrast}

\item{correction}{correction for parameters, none, Occam's bonus (default)
or AIC}

\item{plt}{\code{TRUE} or \code{FALSE} (default), provide interaction plot}

\item{ME}{\code{TRUE} or \code{FALSE} (default), for showing main effects
model in plot}

\item{text}{\code{TRUE} (default) or \code{FALSE}, how to report the
results}

\item{formula}{(optional) the formula to use, see the examples}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$anova} \tab \tab \tab \tab \tab a table of the test results \cr
  \code{results$cvals} \tab \tab \tab \tab \tab a table of the test results \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$tabText} \tab \tab \tab \tab \tab a html \cr
  \code{results$SupportTab} \tab \tab \tab \tab \tab a table \cr
  \code{results$MoretabText} \tab \tab \tab \tab \tab a html \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$anova$asDF}

\code{as.data.frame(results$anova)}
}
\description{
The Analysis of Variance (ANOVA) is used to explore the relationship 
between a continuous dependent variable and two categorical explanatory 
variables. Contrasts can be specified. Support values can be corrected 
according to the number of parameters in the models being compared.
}
\examples{
data('ToothGrowth')
jeva::l2waov(data = ToothGrowth,
dep = len,
factor1 = dose,
factor2 = supp,
text = FALSE)
#
# TWO-WAY FACTORIAL ANOVA
#
# Support
# -----------------------------------------------------------------------------------------------
#   Model Comparisons                 S               Param    F            df       p
# -----------------------------------------------------------------------------------------------
#   H₀ vs full model                  -44.85628744     2, 7    41.557178    5, 54    < .0000001
#   H₀ versus dose                     35.40697964     4, 2    91.999965    2, 54    < .0000001
#   H₀ versus supp                      1.33978727     3, 2    15.571979    1, 54     0.0002312
#   H₀ versus both main effects        41.60841579     5, 2    66.523970    3, 54    < .0000001
#   H₀ versus Interaction              -0.04361461     4, 2     4.106991    2, 54     0.0218603
#   Full model versus main effects      3.24787166     7, 5
#   Full model versus interaction      44.89990206     7, 4
#   H₀ versus Contrast 1
#   H₀ versus Contrast 2
#   Contrast 1 versus Contrast 2
# -----------------------------------------------------------------------------------------------
#   Note. S uses Occam's Bonus correction for parameters (Param).
#

}
