% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laov1w.h.R
\name{laov1w}
\alias{laov1w}
\title{One-Way ANOVA}
\usage{
laov1w(
  data,
  dep,
  group,
  lint = 2,
  ct1 = FALSE,
  Ct1Values = "1,-1,...",
  ct2 = FALSE,
  Ct2Values = "1,-1,...",
  correction = "ob",
  desc = FALSE,
  descPlot = FALSE,
  text = TRUE,
  norm = FALSE,
  qq = FALSE,
  eqv = FALSE,
  formula
)
}
\arguments{
\item{data}{the data as a data frame}

\item{dep}{a string naming the dependent variables in \code{data}}

\item{group}{the grouping or independent variable in \code{data}}

\item{lint}{likelihood interval given as support values, e.g. 2 or 3,
default = 2}

\item{ct1}{\code{TRUE} or \code{FALSE} (default), for contrast 1}

\item{Ct1Values}{a comma-separated list specifying the contrast}

\item{ct2}{\code{TRUE} or \code{FALSE} (default), for contrast 2}

\item{Ct2Values}{a comma-separated list specifying the contrast}

\item{correction}{correction for parameters, none, Occam's bonus (default)
or AIC}

\item{desc}{\code{TRUE} or \code{FALSE} (default), provide descriptive
statistics}

\item{descPlot}{\code{TRUE} or \code{FALSE} (default), provide descriptive
plots}

\item{text}{\code{TRUE} (default) or \code{FALSE}, how to report the
results}

\item{norm}{\code{TRUE} or \code{FALSE} (default), perform Shapiro-Wilk
test of normality}

\item{qq}{\code{TRUE} or \code{FALSE} (default), provide a Q-Q plot of
residuals}

\item{eqv}{\code{TRUE} or \code{FALSE} (default), perform Levene's test for
homogeneity of variances}

\item{formula}{(optional) the formula to use, see the examples}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$anova} \tab \tab \tab \tab \tab a table of the test results \cr
  \code{results$desc} \tab \tab \tab \tab \tab a table containing the group descriptives \cr
  \code{results$assump$eqv} \tab \tab \tab \tab \tab a table of homogeneity of variances tests \cr
  \code{results$assump$norm} \tab \tab \tab \tab \tab a table containing the normality tests \cr
  \code{results$plots} \tab \tab \tab \tab \tab an array of groups of plots \cr
  \code{results$tabText} \tab \tab \tab \tab \tab a html \cr
  \code{results$SupportTab} \tab \tab \tab \tab \tab a table \cr
  \code{results$MoretabText} \tab \tab \tab \tab \tab a html \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$anova$asDF}

\code{as.data.frame(results$anova)}
}
\description{
The Analysis of Variance (ANOVA) is used to explore the relationship
between a continuous dependent variable, and one or more categorical
explanatory variables. This 'One-Way ANOVA' is a simplified version of
the 'normal' ANOVA, allowing only a single explanatory factor and.
dependent variables. Contrasts can be specified whose weights should sum
to zero, but do not need to be orthogonal to each other. Support values, 
S, represent log likelihood ratio. They are given for each model 
comparison. Sc represent Akaike corrected values according to the 
number of parameters
}
\examples{
data('ToothGrowth')
jeva::laov1w(formula = len ~ dose, data = ToothGrowth, text = FALSE)

#
# ONE-WAY ANOVA
#
# Support: One-Way ANOVA (Fisher's assuming equal variances)
# ---------------------------------------------------------------------------------------------------------
#   Model Comparisons                           S              Param    F             df       p
# ---------------------------------------------------------------------------------------------------------
#   H₀  vs observed means                       -35.4069796     2, 4     67.415738    2, 57    < .0000001
#   Contrast 1 (Linear) vs observed means        -0.2362044     3, 4    133.415383    1, 57    < .0000001
#   Contrast 2 (Quadratic) vs observed means    -35.6846993     3, 4      1.416093    1, 57     0.2389805
#   Contrast 1 vs Contrast 2                     35.4484948     3, 3                   1, 1
#   Linear vs Non-linear                         35.4484948     3, 3                   1, 1
# ---------------------------------------------------------------------------------------------------------
#   Note. S uses Occam's Bonus correction for parameters (Param).  Unless specified, Contrast 1 is
#   linear and Contrast 2 is quadratic
#

}
