% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcorr.h.R
\name{lcorr}
\alias{lcorr}
\title{Correlation}
\usage{
lcorr(
  data,
  depa,
  depb,
  alt = 0,
  lint = 2,
  ciWidth = 95,
  correction = "ob",
  pll = FALSE,
  plotype = "lplot",
  supplot = -10,
  plt = FALSE,
  line = "none",
  se = FALSE,
  text = TRUE
)
}
\arguments{
\item{data}{the data as a data frame}

\item{depa}{the first variable from \code{data}, plotted on vertical axis,
variable must be numeric}

\item{depb}{the second variable from \code{data}, plotted on horizontal
axis, variable must be numeric}

\item{alt}{a specified value for the alternative hypothesis, range -1 to
+1, default 0}

\item{lint}{likelihood interval given as support value, e.g. 2 or 3,
default = 2}

\item{ciWidth}{a number between 50 and 99.9 (default: 95), width of the
confidence intervals to provide}

\item{correction}{correction for parameters, none, Occam's bonus (default)
or AIC}

\item{pll}{plot the likelihood function displaying observed correlation,
alternative hypothesis and support interval}

\item{plotype}{choose type of plot, likelihood function (default), support
function}

\item{supplot}{To set the minimum likelihood display value in plot, in log
units (default = -10)  affects the x-axis range}

\item{plt}{\code{TRUE} or \code{FALSE} (default), do a scatter plot of the
data}

\item{line}{\code{none} (default), \code{linear}, or \code{smooth}, provide
respectively no regression line,  a linear regression line, or a smoothed
regression line}

\item{se}{\code{TRUE} or \code{FALSE} (default), show the standard error
for the regression line}

\item{text}{\code{TRUE} (default) or \code{FALSE}, how to report the
results}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$lcor} \tab \tab \tab \tab \tab a table \cr
  \code{results$lcor1} \tab \tab \tab \tab \tab a table \cr
  \code{results$plotc} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$tabText} \tab \tab \tab \tab \tab a html \cr
  \code{results$SupportTab} \tab \tab \tab \tab \tab a table \cr
  \code{results$MoretabText} \tab \tab \tab \tab \tab a html \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$lcor$asDF}

\code{as.data.frame(results$lcor)}
}
\description{
Bivariate correlation analysis where an alternative hypothesis correlation 
can be specified. Likelihood support and likelihood based confidence 
intervals can be specified.
}
\examples{
data('mtcars')
jeva::lcorr(data = mtcars, depa = mpg, depb = cyl, text = FALSE)

#
# CORRELATION
#
# Support
#                                     Value       Difference    S              Param    t            df    p             z            p(z)
# ------------------------------------------------------------------------------------------------------------------------------------------------
#   H₀ vs observed correlation        0.000000     0.8521620    -22.6664364     2, 3    -8.919699    30    < .0000001    -6.806825    < .0000001
#   Ha versus observed correlation    0.000000     0.8521620    -23.1664364     3, 3                                     -6.806825    < .0000001
#   Ha vs H₀                                       0.0000000     -0.5000000     3, 2
# ------------------------------------------------------------------------------------------------------------------------------------------------
#   Note. S uses Occam's Bonus correction for parameters (Param). The last column p(z) gives the p value for the preceding z value
#
# Intervals
# -------------------------------------------------------------------------
#   Type of interval      Level    r             Lower         Upper
# -------------------------------------------------------------------------
#   Support               2        -0.8521620    -0.9268248    -0.7126780
#   Likelihood-based a    95\\%      -0.8521620    -0.9257632    -0.7163143
# -------------------------------------------------------------------------
#
#

}
