% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldiag.h.R
\name{ldiag}
\alias{ldiag}
\title{Diagnostic Test}
\usage{
ldiag(
  data,
  rows,
  cols,
  counts = NULL,
  nulsens = 0.5,
  nulspec = 0.5,
  pprob = FALSE,
  ppval = 0.1,
  lint = 2,
  ciWidth = 95,
  correction = "ob",
  fagan = FALSE,
  pll = FALSE,
  plotype = "lplot",
  supplot = -10,
  choice = "sens_plot",
  varA = FALSE,
  cc = FALSE,
  text = TRUE,
  obs = TRUE,
  exp = FALSE,
  pcRow = FALSE,
  pcCol = FALSE,
  pcTot = FALSE,
  barplot = FALSE,
  yaxis = "ycounts",
  yaxisPc = "total_pc",
  xaxis = "xrows",
  bartype = "dodge",
  formula
)
}
\arguments{
\item{data}{the data as a data frame}

\item{rows}{the variable to use as the rows in the contingency table (not
necessary when providing a formula, see the examples)}

\item{cols}{the variable to use as the columns in the contingency table
(not necessary when providing a formula, see the examples)}

\item{counts}{the variable to use as the counts in the contingency table
(not necessary when providing a formula, see the examples)}

\item{nulsens}{value for the null hypothesis, default = 0.5}

\item{nulspec}{value for the alternative hypothesis, default = 0.5}

\item{pprob}{\code{TRUE} or \code{FALSE} (default), select the prior
probability or prevalence}

\item{ppval}{To set the prior probability/prevalence}

\item{lint}{likelihood interval given as support value, e.g. 2 or 3,
default = 2}

\item{ciWidth}{a number between 50 and 99.9 (default: 95), width of the
confidence intervals to provide}

\item{correction}{correction for parameters, none, Occam's bonus (default)
or AIC}

\item{fagan}{.}

\item{pll}{\code{TRUE} or \code{FALSE} (default), give the likelihood
function showing null hypothesis (black line),  alternative hypothesis
(blue line), mean (dashed line), and specified support interval
(horizontal red line)}

\item{plotype}{choose type of plot, likelihood function (default), support
function}

\item{supplot}{To set the minimum likelihood display value in plot, in log
units (default = -10)  affects the x-axis range}

\item{choice}{choose to show likelihood function for sensitivity (default)
or specificity}

\item{varA}{\code{TRUE} or \code{FALSE} (default), perform variance
analysis for null and alternative hypothesis}

\item{cc}{\code{TRUE} or \code{FALSE} (default), use continuity correction}

\item{text}{\code{TRUE} (default) or \code{FALSE}, how to report the
results}

\item{obs}{\code{TRUE} or \code{FALSE} (default), provide the observed
counts}

\item{exp}{\code{TRUE} or \code{FALSE} (default), provide the expected
counts}

\item{pcRow}{\code{TRUE} or \code{FALSE} (default), provide row percentages}

\item{pcCol}{\code{TRUE} or \code{FALSE} (default), provide column
percentages}

\item{pcTot}{\code{TRUE} or \code{FALSE} (default), provide total
percentages}

\item{barplot}{\code{TRUE} or \code{FALSE} (default), show barplots}

\item{yaxis}{ycounts (default) or ypc. Use respectively \code{counts} or
\code{percentages} for the bar plot y-axis}

\item{yaxisPc}{total_pc (default), column_pc, or row_pc. Use respectively
percentages \code{of total}, \code{within columns}, or \code{within rows}
for the bar plot y-axis.}

\item{xaxis}{rows (default), or columns in bar plot X axis}

\item{bartype}{stack or side by side (default), barplot type}

\item{formula}{(optional) the formula to use, see the examples}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$freqs} \tab \tab \tab \tab \tab a table of proportions \cr
  \code{results$stats} \tab \tab \tab \tab \tab a table \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$ctt} \tab \tab \tab \tab \tab a table \cr
  \code{results$cttsens} \tab \tab \tab \tab \tab a table \cr
  \code{results$cttspec} \tab \tab \tab \tab \tab a table \cr
  \code{results$ctt3} \tab \tab \tab \tab \tab a table \cr
  \code{results$plotc} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$barplot} \tab \tab \tab \tab \tab an image \cr
  \code{results$tabText} \tab \tab \tab \tab \tab a html \cr
  \code{results$SupportTab} \tab \tab \tab \tab \tab a table \cr
  \code{results$MoretabText} \tab \tab \tab \tab \tab a html \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$freqs$asDF}

\code{as.data.frame(results$freqs)}
}
\description{
Analysis of a 2 x 2 contingency table using diagnostic test
}
\examples{
dat <- data.frame(
  CT_scan = c('+', '+', '-', '-'),
  Final_diagnosis = c('+', '-','+', '-'),
  count = c(52, 11, 37, 100),
  check.names=FALSE)

  jeva::ldiag(formula = count ~ CT_scan:Final_diagnosis, data = dat, text = FALSE)

}
