% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lreg.h.R
\name{lreg}
\alias{lreg}
\title{Polynomial Regression}
\usage{
lreg(
  data,
  dep,
  pred,
  correction = "ob",
  plt = FALSE,
  lin = TRUE,
  quad = FALSE,
  cub = FALSE,
  quart = FALSE,
  lint = 2,
  text = TRUE
)
}
\arguments{
\item{data}{the data as a data frame}

\item{dep}{the dependent variable from \code{data}, variable must be
numeric}

\item{pred}{the predictor variable, must be numeric}

\item{correction}{correction for parameters, none, Occam's bonus (default)
or AIC}

\item{plt}{\code{TRUE} or \code{FALSE} (default), do a scatter plot of the
data}

\item{lin}{\code{TRUE} (default) or \code{FALSE}, plot line for linear
relationship}

\item{quad}{\code{TRUE} or \code{FALSE} (default), plot curve for quadratic
relationship}

\item{cub}{\code{TRUE} or \code{FALSE} (default), plot curve for cubic
relationship}

\item{quart}{\code{TRUE} or \code{FALSE} (default), plot curve for quartic
relationship}

\item{lint}{a number between 1 and 100 (default: 2) specifying the
likelihood support interval width}

\item{text}{\code{TRUE} (default) or \code{FALSE}, how to report the
results}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$lreg} \tab \tab \tab \tab \tab a table \cr
  \code{results$coef} \tab \tab \tab \tab \tab a table of model coefficients \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$tabText} \tab \tab \tab \tab \tab a html \cr
  \code{results$SupportTab} \tab \tab \tab \tab \tab a table \cr
  \code{results$MoretabText} \tab \tab \tab \tab \tab a html \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$lreg$asDF}

\code{as.data.frame(results$lreg)}
}
\description{
Polynomial regression of two variables where different orders are compared. 
Akaike corrected support values are calculated hierarchically. R-squared 
and AIC values are given. P values are calculated in the usual way from the 
residual mean sums of squares from the highest polynomial in the model, but 
these are not comparable with the support values. Model coefficients are 
given with their specified likelihood support level. Plots for different 
polynomial orders can be plotted.
}
\examples{
data('mtcars')
jeva::lreg(data = mtcars, dep = mpg, pred = hp, text = FALSE)

#
# POLYNOMIAL REGRESSION
#
# Support
# --------------------------------------------------------------------------------------------------
#   Model Comparisons          S              Param    R²           AIC         df       p
# --------------------------------------------------------------------------------------------------
#   H₀ vs Linear               -14.2584436     2, 3    0.6024373    181.2386    1, 30    0.0000002
#   Quadratic vs Linear          7.3181632     3, 4    0.7561100    167.6023    1, 27    0.0002868
#   Cubic vs Lower Orders       -0.2044684     4, 5    0.7605735    169.0112    1, 27    0.4840465
#   Quartic vs Lower Orders     -0.4948111     5, 6    0.7606511    171.0009    1, 27    0.9261317
# --------------------------------------------------------------------------------------------------
#   Note. S uses Occam's Bonus correction for parameters (Param). S is calculated hierarchically
#   from all remaining SSq, while p values are calculated using residual MSq from the highest
#   polynomial model.
#
#
# Model Coefficients - mpg
# ----------------------------------------------------------------------------------------------------------
#   Predictor    Estimate        SE              S               Lower           Upper           p
# ----------------------------------------------------------------------------------------------------------
#   Intercept    45.359415905    13.557249652    -5.549534400    19.654183298    71.064648512    0.0024227
#   Linear       -0.328488135     0.382186016    -0.431888228    -1.053132196     0.396155925    0.3976331
#   Quadratic     0.001246952     0.003652374    -0.068923891    -0.005678134     0.008172038    0.7354387
#   Cubic        -2.185873e-6     1.411330e-5    -0.014208762    -2.894540e-5     2.457366e-5    0.8780676
#   Quartic       1.756358e-9     1.876800e-8    -0.005188909    -3.382872e-8     3.734143e-8    0.9261317
# ----------------------------------------------------------------------------------------------------------
#

}
