% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrm1waov.h.R
\name{lrm1waov}
\alias{lrm1waov}
\title{Repeated Measures ANOVA}
\usage{
lrm1waov(
  data,
  rm = list(list(label = "RM Factor 1", levels = list("Level 1", "Level 2"))),
  rmCells = NULL,
  bs = NULL,
  ct1 = FALSE,
  Ct1Values = "1,-1,...",
  ct2 = FALSE,
  Ct2Values = "1,-1,...",
  correction = "ob",
  effectSize = NULL,
  depLabel = "Dependent",
  rmTerms = NULL,
  bsTerms = NULL,
  ss = "3",
  spherTests = FALSE,
  spherCorr = list("none"),
  leveneTest = FALSE,
  qq = FALSE,
  contrasts = NULL,
  emMeans = list(list()),
  emmPlots = TRUE,
  emmTables = FALSE,
  emmWeights = TRUE,
  ciWidthEmm = 95,
  emmPlotData = FALSE,
  emmPlotError = "ci",
  groupSumm = FALSE,
  text = TRUE
)
}
\arguments{
\item{data}{the data as a data frame}

\item{rm}{a list of lists, where each list describes the \code{label} (as a
string) and the \code{levels} (as vector of strings) of a particular
repeated measures factor}

\item{rmCells}{a list of lists, where each list decribes a repeated measure
(as a string) from \code{data} defined as \code{measure} and the particular
combination of levels from \code{rm} that it belongs to (as a vector of
strings) defined as \code{cell}}

\item{bs}{a string naming the between subjects factor from \code{data}}

\item{ct1}{\code{TRUE} or \code{FALSE} (default), for contrast 1}

\item{Ct1Values}{a comma-separated list specifying the contrast}

\item{ct2}{\code{TRUE} or \code{FALSE} (default), for contrast 2}

\item{Ct2Values}{a comma-separated list specifying the contrast}

\item{correction}{correction for parameters, none, Occam's bonus (default)
or AIC}

\item{effectSize}{one or more of \code{'eta'}, \code{'partEta'}, or
\code{'omega'}; use eta², partial eta², and omega² effect sizes,
respectively}

\item{depLabel}{a string (default: 'Dependent') describing the label used
for the dependent variable throughout the analysis}

\item{rmTerms}{a list of character vectors describing the repeated measures
terms to go into the model}

\item{bsTerms}{a list of character vectors describing the between subjects
terms to go into the model}

\item{ss}{\code{'2'} or \code{'3'} (default), the sum of squares to use}

\item{spherTests}{\code{TRUE} or \code{FALSE} (default), perform sphericity
tests}

\item{spherCorr}{one or more of \code{'none'} (default), \code{'GG'}, or
\code{}HF\code{}; use no p-value correction, the Greenhouse-Geisser p-value
correction, and the Huynh-Feldt p-value correction for shericity,
respectively}

\item{leveneTest}{\code{TRUE} or \code{FALSE} (default), test for
homogeneity of variances (i.e., Levene's test)}

\item{qq}{\code{TRUE} or \code{FALSE} (default), provide a Q-Q plot of
residuals}

\item{contrasts}{in development}

\item{emMeans}{a list of lists specifying the variables for which the
estimated marginal means need to be calculate. Supports up to three
variables per term.}

\item{emmPlots}{\code{TRUE} (default) or \code{FALSE}, provide estimated
marginal means plots}

\item{emmTables}{\code{TRUE} or \code{FALSE} (default), provide estimated
marginal means tables}

\item{emmWeights}{\code{TRUE} (default) or \code{FALSE}, weigh each cell
equally or weigh them according to the cell frequency}

\item{ciWidthEmm}{a number between 50 and 99.9 (default: 95) specifying the
confidence interval width for the estimated marginal means}

\item{emmPlotData}{\code{TRUE} or \code{FALSE} (default), plot the data on
top of the marginal means}

\item{emmPlotError}{\code{'none'}, \code{'ci'} (default), or \code{'se'}.
Use no error bars, use confidence intervals, or use standard errors on the
marginal mean plots, respectively}

\item{groupSumm}{\code{TRUE} or \code{FALSE} (default), report a summary of
the different groups}

\item{text}{\code{TRUE} (default) or \code{FALSE}, how to report the
results}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$rmTable} \tab \tab \tab \tab \tab a table \cr
  \code{results$bsTable} \tab \tab \tab \tab \tab a table \cr
  \code{results$conts} \tab \tab \tab \tab \tab a table of the test results \cr
  \code{results$cvals} \tab \tab \tab \tab \tab a table of the test results \cr
  \code{results$assump$spherTable} \tab \tab \tab \tab \tab a table \cr
  \code{results$assump$leveneTable} \tab \tab \tab \tab \tab a table \cr
  \code{results$assump$qq} \tab \tab \tab \tab \tab a q-q plot \cr
  \code{results$contrasts} \tab \tab \tab \tab \tab an array of tables \cr
  \code{results$emm} \tab \tab \tab \tab \tab an array of the estimated marginal means plots + tables \cr
  \code{results$groupSummary} \tab \tab \tab \tab \tab a summary of the groups \cr
  \code{results$tabText} \tab \tab \tab \tab \tab a html \cr
  \code{results$SupportTab} \tab \tab \tab \tab \tab a table \cr
  \code{results$MoretabText} \tab \tab \tab \tab \tab a html \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$rmTable$asDF}

\code{as.data.frame(results$rmTable)}
}
\description{
The Repeated Measures ANOVA is used to explore the relationship
between a continuous dependent variable and one or two categorical
explanatory variables, where one or more of the explanatory variables
are 'within subjects' (where multiple measurements are from the same
subject).
}
\details{
This analysis requires that the data be in 'wide format', where each
row represents a subject (as opposed to long format, where each
measurement of the dependent variable is represented as a row).

Currently, this log likelihood analysis does not accept covariates.
}
\examples{

data('bugs', package = 'jmv')

jeva::lrm1waov(
    data = bugs,
    rm = list(
        list(
            label = 'Frightening',
            levels = c('Low', 'High'))),
    rmCells = list(
        list(
            measure = 'LDLF',
            cell = 'Low'),
        list(
            measure = 'LDHF',
            cell = 'High')),
    rmTerms = list(
        'Frightening'))

# REPEATED MEASURES ANOVA
#
#   Within Subjects Effects
#   -------------------------------------------------------------------------------
#                           S            Param    F           df       p
#   -------------------------------------------------------------------------------
#     H₀  vs Frightening    -17.67963     3, 4    44.20412    1, 90    < .0000001
#   -------------------------------------------------------------------------------
#     Note. Type 3 Sums of Squares.  S uses Occam's Bonus correction for
#     parameters (Param).
#

}
