% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrttest.h.R
\name{lrttest}
\alias{lrttest}
\title{Paired Samples T-Test}
\usage{
lrttest(
  data,
  depa,
  depb,
  nul = 0,
  alt = 0,
  lint = 2,
  correction = "ob",
  pll = FALSE,
  plotype = "lplot",
  supplot = -10,
  dtab = FALSE,
  plt = FALSE,
  text = TRUE
)
}
\arguments{
\item{data}{the data as a data frame}

\item{depa}{a (non-empty) numeric vector of data values for 1st group}

\item{depb}{a (non-empty) numeric vector of data values for 2nd group}

\item{nul}{difference value for the null hypothesis, default = 0}

\item{alt}{difference value for an alternative hypothesis, default = 0}

\item{lint}{likelihood interval given as support value, e.g. 2 or 3,
default = 2}

\item{correction}{correction for parameters, none, Occam's bonus (default)
or AIC}

\item{pll}{\code{TRUE} or \code{FALSE} (default), give the likelihood or
log likelihood function showing  null hypothesis (black line), alternative
hypothesis (blue line), mean (dashed line),  and specified support interval
(horizontal red line)}

\item{plotype}{choose type of plot, likelihood function (default), support
function}

\item{supplot}{To set the minimum likelihood display value in plot, in log
units (default = -10)  affects the x-axis range}

\item{dtab}{\code{TRUE} or \code{FALSE} (default), provide descriptive
statistics}

\item{plt}{\code{TRUE} or \code{FALSE} (default), give a plot of means with
specified support intervals}

\item{text}{\code{TRUE} (default) or \code{FALSE}, how to report the
results}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$lrttest} \tab \tab \tab \tab \tab a table \cr
  \code{results$lrttest2} \tab \tab \tab \tab \tab a table \cr
  \code{results$lrttestd} \tab \tab \tab \tab \tab a table \cr
  \code{results$tabText} \tab \tab \tab \tab \tab a html \cr
  \code{results$SupportTab} \tab \tab \tab \tab \tab a table \cr
  \code{results$MoretabText} \tab \tab \tab \tab \tab a html \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plotc} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$lrttest$asDF}

\code{as.data.frame(results$lrttest)}
}
\description{
Related samples analysis where null and alternative hypotheses can be 
specified. Likelihood interval support level can be specified and plotted 
with the mean difference. A likelihood function for the mean difference can 
be plotted.
}
\examples{
data('bugs', package = 'jmv')
jeva::lrttest(data = bugs, depa = LDLF, depb = LDHF, text = FALSE)

#
# PAIRED SAMPLES T-TEST
#
# Support
#                                     Value       Difference    SE difference    S              Param    t            df          p
# --------------------------------------------------------------------------------------------------------------------------------------------
#   H₀ vs observed mean difference    0.000000      1.664835        0.2504032    -17.6796259     2, 3    -6.648618    90.00000    < .0000001
#   Ha vs observed mean difference    0.000000      1.664835        0.2504032    -18.1796259     3, 3    -6.648618    90.00000    < .0000001
#   Ha vs H₀                                        0.000000        0.2504032     -0.5000000     3, 2
# --------------------------------------------------------------------------------------------------------------------------------------------
#   Note. S uses Occam's Bonus correction for parameters (Param)
#
#
# Support Interval for mean difference
# ---------------------------------------------
#   Mean difference    Lower        Upper
# ---------------------------------------------
#         -1.664835    -2.168406    -1.161265
# ---------------------------------------------
#

}
