% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsampsz.h.R
\name{lsampsz}
\alias{lsampsz}
\title{Sample Size Calculator}
\usage{
lsampsz(
  ttype = "is",
  MW = 0.05,
  S = 3,
  d = 1.2,
  plwm = FALSE,
  tail1 = "onet",
  alpha = 0.05,
  text = TRUE
)
}
\arguments{
\item{ttype}{type of t-test, independent (default) or paired/one sample}

\item{MW}{probability of misleading + weak evidence, default = 0.05}

\item{S}{strength of evidence required LLR, default = 3}

\item{d}{Cohen's d effect size, default = 1.2}

\item{plwm}{\code{TRUE} or \code{FALSE} (default), show the probability
plots for weak  and misleading evidence}

\item{tail1}{whether one- (default) or two-tailed for beta calculations}

\item{alpha}{probability of Type I error, default = 0.05}

\item{text}{\code{TRUE} (default) or \code{FALSE}, brief explanation of
evidential sample size analysis}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$tab} \tab \tab \tab \tab \tab a table \cr
  \code{results$space1} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$space2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plotp} \tab \tab \tab \tab \tab an image \cr
  \code{results$plotText} \tab \tab \tab \tab \tab a html \cr
  \code{results$exptext} \tab \tab \tab \tab \tab a html \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$tab$asDF}

\code{as.data.frame(results$tab)}
}
\description{
Sample size calculation for all t-tests, specifying strength of evidence 
required, the probability of misleading and weak evidence, and the effect 
size. Shows a probability plot of types of error against sample size.
}
\examples{
jeva::lsampsz(ttype = "is", MW = 0.05, S = 3, d = 1.2, text = FALSE)

#
# SAMPLE SIZE CALCULATOR
#
# Evidential sample size estimate
# ---------------------------------------------------------------------------------
#                               MW prob       S           d           Sample size
# ---------------------------------------------------------------------------------
#   Independent samples test    0.05000000    3.000000    1.200000       42.00000
# ---------------------------------------------------------------------------------
#   Note. Divide the given sample size by 2 to get the number for each group
#

}
