% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lttest.h.R
\name{lttest}
\alias{lttest}
\title{One Sample T-Test}
\usage{
lttest(
  data,
  dep,
  nul = 0,
  alt = 0,
  lint = 2,
  correction = "ob",
  pll = FALSE,
  plotype = "lplot",
  supplot = -10,
  dtab = FALSE,
  plt = FALSE,
  text = TRUE
)
}
\arguments{
\item{data}{the data as a data frame}

\item{dep}{a (non-empty) numeric vector of data values}

\item{nul}{value for the null hypothesis, default = 0}

\item{alt}{value for an alternative hypothesis, default = 0}

\item{lint}{likelihood interval given as support values, e.g. 2 or 3,
default = 2}

\item{correction}{correction for parameters, none, Occam's bonus (default)
or AIC}

\item{pll}{\code{TRUE} or \code{FALSE} (default), give the likelihood or
log likelihood function showing  null hypothesis (black line), alternative
hypothesis (blue line), mean (dashed line),  and specified support interval
(horizontal red line)}

\item{plotype}{choose type of plot, likelihood function (default), support
function}

\item{supplot}{To set the minimum likelihood display value in plot, in log
units (default = -10)  affects the x-axis range}

\item{dtab}{\code{TRUE} or \code{FALSE} (default), provide descriptive
statistics}

\item{plt}{\code{TRUE} or \code{FALSE} (default), give a plot of means with
specified support intervals}

\item{text}{\code{TRUE} (default) or \code{FALSE}, how to report the
results}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$lttest} \tab \tab \tab \tab \tab a table \cr
  \code{results$lttest2} \tab \tab \tab \tab \tab a table \cr
  \code{results$lttestd} \tab \tab \tab \tab \tab a table \cr
  \code{results$tabText} \tab \tab \tab \tab \tab a html \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plotc} \tab \tab \tab \tab \tab an image \cr
  \code{results$SupportTab} \tab \tab \tab \tab \tab a table \cr
  \code{results$MoretabText} \tab \tab \tab \tab \tab a html \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$lttest$asDF}

\code{as.data.frame(results$lttest)}
}
\description{
One sample analysis where null and alternative hypotheses can be specified. 
Likelihood interval support level can be specified and plotted with the 
mean. A likelihood function for the mean can be plotted.
}
\examples{
data('ToothGrowth')
jeva::lttest(data = ToothGrowth, dep = len, text = FALSE)

#
# ONE SAMPLE T-TEST
#
# Support
# ----------------------------------------------------------------------------------------------------------------------
#                          Value       Difference    SE           S              Param    t           df    p
# ----------------------------------------------------------------------------------------------------------------------
#   H₀ vs observed mean    0.000000    -18.813333    0.9875223    -58.5199371     2, 3    19.05105    59    < .0000001
#   Ha vs observed mean    0.000000    -18.813333    0.9875223    -59.0199371     3, 3    19.05105    59    < .0000001
#   Ha vs H₀                             0.000000    0.9875223     -0.5000000     3, 2
# ----------------------------------------------------------------------------------------------------------------------
#   Note. S uses Occam's Bonus correction for parameters (Param).
#
# Support Interval for mean
# ------------------------------------
#   Mean        Lower       Upper
# ------------------------------------
#   18.81333    16.82172    20.80495
# ------------------------------------
#

}
