% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnem.h.R
\name{mcnem}
\alias{mcnem}
\title{Paired Samples}
\usage{
mcnem(
  data,
  rows,
  cols,
  counts = NULL,
  nul = 1,
  alt = 1,
  lint = 2,
  ciWidth = 95,
  correction = "ob",
  pll = FALSE,
  plotype = "lplot",
  supplot = -10,
  obs = TRUE,
  exp = FALSE,
  pcRow = FALSE,
  pcCol = FALSE,
  pcTot = FALSE,
  varA = FALSE,
  text = TRUE,
  barplot = FALSE,
  yaxis = "ycounts",
  yaxisPc = "total_pc",
  xaxis = "xrows",
  bartype = "dodge",
  formula
)
}
\arguments{
\item{data}{the data as a data frame}

\item{rows}{the variable to use as the rows in the contingency table}

\item{cols}{the variable to use as the columns in the contingency table}

\item{counts}{the variable to use as the counts in the contingency table
(not necessary when providing a formula, see the examples)}

\item{nul}{value for the null hypothesis, default = 1}

\item{alt}{value for the alternative hypothesis, default = 1}

\item{lint}{likelihood interval given as support value, e.g. 2 or 3,
default = 2}

\item{ciWidth}{a number between 50 and 99.9 (default: 95), width of the
confidence intervals to provide}

\item{correction}{correction for parameters, none, Occam's bonus (default)
or AIC}

\item{pll}{\code{TRUE} or \code{FALSE} (default), give the likelihood
function showing null hypothesis (black line),  alternative hypothesis
(blue line), mean (dashed line), and specified support interval
(horizontal red line)}

\item{plotype}{choose type of plot, likelihood function (default), support
function}

\item{supplot}{To set the minimum likelihood display value in plot, in log
units (default = -10)  affects the x-axis range}

\item{obs}{\code{TRUE} or \code{FALSE} (default), provide the observed
counts}

\item{exp}{\code{TRUE} or \code{FALSE} (default), provide the expected
counts}

\item{pcRow}{\code{TRUE} or \code{FALSE} (default), provide row percentages}

\item{pcCol}{\code{TRUE} or \code{FALSE} (default), provide column
percentages}

\item{pcTot}{\code{TRUE} or \code{FALSE} (default), provide total
percentages}

\item{varA}{\code{TRUE} or \code{FALSE} (default), perform variance
analysis for null and alternative hypotheses}

\item{text}{\code{TRUE} (default) or \code{FALSE}, how to report the
results}

\item{barplot}{\code{TRUE} or \code{FALSE} (default), show barplots}

\item{yaxis}{ycounts (default) or ypc. Use respectively \code{counts} or
\code{percentages} for the bar plot y-axis}

\item{yaxisPc}{total_pc (default), column_pc, or row_pc. Use respectively
percentages \code{of total}, \code{within columns}, or \code{within rows}
for the bar plot y-axis.}

\item{xaxis}{rows (default), or columns in bar plot X axis}

\item{bartype}{stack or side by side (default), barplot type}

\item{formula}{(optional) the formula to use, see the examples}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$freqs} \tab \tab \tab \tab \tab a table of proportions \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$ctt} \tab \tab \tab \tab \tab a table \cr
  \code{results$ctt2} \tab \tab \tab \tab \tab a table \cr
  \code{results$tabText} \tab \tab \tab \tab \tab a html \cr
  \code{results$SupportTab} \tab \tab \tab \tab \tab a table \cr
  \code{results$MoretabText} \tab \tab \tab \tab \tab a html \cr
  \code{results$ctt3} \tab \tab \tab \tab \tab a table \cr
  \code{results$plotc} \tab \tab \tab \tab \tab an image \cr
  \code{results$barplot} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$freqs$asDF}

\code{as.data.frame(results$freqs)}
}
\description{
Paired samples analysis (McNemar test) where null and alternative odds can 
be specified.
}
\examples{
dat <- data.frame(
First = c('Approve', 'Approve', 'Disapprove', 'Disapprove'),
Second = c('Approve', 'Disapprove', 'Approve', 'Disapprove'),
Counts = c(794, 150, 86, 570))

jeva::mcnem(formula = Counts ~ First:Second, data = dat, text=FALSE)

  #
  # PAIRED SAMPLES - MCNEMAR TEST
  #
  # Contingency Tables
  # ------------------------------------------------
  #   First         Approve    Disapprove    Total
  # ------------------------------------------------
  #   Approve           794           150      944
  #   Disapprove         86           570      656
  #   Total             880           720     1600
  # ------------------------------------------------
  #
  #
  # Support
  # ----------------------------------------------------------------------------------------------------------
  #                          Odds Value    Difference    S             Param    G            df    p
  # ----------------------------------------------------------------------------------------------------------
  #   H₀ vs observed Odds      1.000000     0.4266667    -8.2875901     1, 2    16.575180     1    0.0000468
  #   Ha vs observed Odds      1.000000     0.4266667    -8.7875901     2, 2    17.575180     1    0.0000276
  #   Ha vs H₀                              0.0000000    -0.5000000     2, 1     1.000000     1    0.3173105
  # ----------------------------------------------------------------------------------------------------------
  #   Note. S uses Occam's Bonus correction for parameters (Param).
  #
  #
  # Intervals
  # -------------------------------------------------------------------------------------
  #   Type of interval                  Level    Statistic     Lower         Upper
  # -------------------------------------------------------------------------------------
  #   Support for Odds                  2         0.5733333     0.4358143     0.7492639
  #   Likelihood-based for Odds    a    95\\%       0.5733333     0.4382493     0.7452969
  #   Likelihood-based for Logodds      95\\%      -0.5562880    -0.8249674    -0.2939726
  # -------------------------------------------------------------------------------------
  #   a See reference Pritikin et al (2017) likelihood-based intervals are more
  #   accurate and are parameterization-invariant compared to conventional confidence
  #   intervals
  #

}
