% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.h.R
\name{prop}
\alias{prop}
\title{N Outcomes}
\usage{
prop(
  data,
  var,
  counts = NULL,
  expected = FALSE,
  bi = FALSE,
  ciWidth = 95,
  lint = 2,
  pll = FALSE,
  plotype = "lplot",
  supplot = -10,
  correction = "ob",
  ratio = NULL,
  varA = FALSE,
  text = TRUE,
  formula
)
}
\arguments{
\item{data}{the data as a data frame}

\item{var}{categorical or ordinal variable of interest}

\item{counts}{the counts in \code{data}}

\item{expected}{\code{TRUE} or \code{FALSE} (default), whether expected
counts should be displayed}

\item{bi}{\code{TRUE} or \code{FALSE} (default), only if there are 2 levels
of variable}

\item{ciWidth}{a number between 50 and 99.9 (default: 95), width of the
confidence intervals to provide}

\item{lint}{a number between 1 and 100 (default: 2) specifying the
likelihood support interval width}

\item{pll}{for binomial only: plot the likelihood function displaying
observed proportion,  alternative hypothesis and support interval}

\item{plotype}{choose type of plot, likelihood function (default), support
function}

\item{supplot}{To set the minimum likelihood display value in plot, in log
units (default = -10)  affects the x-axis range}

\item{correction}{correction for parameters, none, Occam's bonus (default)
or AIC}

\item{ratio}{a vector of numbers: the expected proportions}

\item{varA}{\code{TRUE} or \code{FALSE} (default), perform variance
analysis for null and alternative hypotheses}

\item{text}{\code{TRUE} (default) or \code{FALSE}, how to report the
results}

\item{formula}{(optional) the formula to use, see the examples}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$props} \tab \tab \tab \tab \tab a table of the proportions \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$tests} \tab \tab \tab \tab \tab a table of the test results \cr
  \code{results$ctt2} \tab \tab \tab \tab \tab a table \cr
  \code{results$tabText} \tab \tab \tab \tab \tab a html \cr
  \code{results$SupportTab} \tab \tab \tab \tab \tab a table \cr
  \code{results$MoretabText} \tab \tab \tab \tab \tab a html \cr
  \code{results$ctt3} \tab \tab \tab \tab \tab a table \cr
  \code{results$plotc} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$props$asDF}

\code{as.data.frame(results$props)}
}
\description{
The one-way analysis examines how well the observed proportions match those 
expected according to a null hypothesis or a specified hypothesis.
}
\examples{
dat <- data.frame(Sex = c('Female', 'Male'), Students = c(60, 40))


jeva::prop(formula = Students ~ Sex, data = dat, ratio = c(1, 1), text = FALSE)

# N OUTCOMES
#
# Proportions - Sex
# ---------------------------------
#   Level     Count    Proportion
# ---------------------------------
#   Female       60     0.6000000
#   Male         40     0.4000000
# ---------------------------------
#
#
# Support
# ----------------------------------------------------------------------------------------------------------
#   Hypotheses                    Expected values    S             Param    G           df    p
# -------------------------------------------------------------------------------------------------------
#   H₀ vs observed proportions    50 | 50            -1.5135514     1, 2    4.027103     1    0.0447748
#   Ha vs observed proportion     50 | 50            -2.0135514     2, 2    4.027103     1    0.0447748
#   Ha vs H₀                                         -0.5000000     2, 1    0.000000     1    1.0000000
# -------------------------------------------------------------------------------------------------------
#   Note. S uses Occam's Bonus correction for parameters (Param).
#

}
