% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llr.h.R
\name{llr}
\alias{llr}
\title{Relative Risk}
\usage{
llr(
  data,
  rows,
  cols,
  counts = NULL,
  nul = 1,
  alt = 1,
  lint = 2,
  ciWidth = 95,
  correction = "ob",
  pll = FALSE,
  plotype = "lplot",
  supplot = -10,
  varA = FALSE,
  cc = FALSE,
  text = TRUE,
  obs = TRUE,
  exp = FALSE,
  pcRow = FALSE,
  pcCol = FALSE,
  pcTot = FALSE,
  barplot = FALSE,
  yaxis = "ycounts",
  yaxisPc = "total_pc",
  xaxis = "xrows",
  bartype = "dodge",
  formula
)
}
\arguments{
\item{data}{the data as a data frame}

\item{rows}{the variable to use as the rows in the contingency table (not
necessary when providing a formula, see the examples)}

\item{cols}{the variable to use as the columns in the contingency table
(not necessary when providing a formula, see the examples)}

\item{counts}{the variable to use as the counts in the contingency table
(not necessary when providing a formula, see the examples)}

\item{nul}{value for the null hypothesis, default = 1}

\item{alt}{value for the alternative hypothesis, default = 1}

\item{lint}{likelihood interval given as support value, e.g. 2 or 3,
default = 2}

\item{ciWidth}{a number between 50 and 99.9 (default: 95), width of the
confidence intervals to provide}

\item{correction}{correction for parameters, none, Occam's bonus (default)
or AIC}

\item{pll}{\code{TRUE} or \code{FALSE} (default), give the likelihood
function showing null hypothesis (black line),  alternative hypothesis
(blue line), mean (dashed line), and specified support interval
(horizontal red line)}

\item{plotype}{choose type of plot, likelihood function (default), support
function}

\item{supplot}{To set the minimum likelihood display value in plot, in log
units (default = -10)  affects the x-axis range}

\item{varA}{\code{TRUE} or \code{FALSE} (default), perform variance
analysis for null and alternative hypothesis}

\item{cc}{\code{TRUE} or \code{FALSE} (default), use continuity correction}

\item{text}{\code{TRUE} (default) or \code{FALSE}, how to report the
results}

\item{obs}{\code{TRUE} or \code{FALSE} (default), provide the observed
counts}

\item{exp}{\code{TRUE} or \code{FALSE} (default), provide the expected
counts}

\item{pcRow}{\code{TRUE} or \code{FALSE} (default), provide row percentages}

\item{pcCol}{\code{TRUE} or \code{FALSE} (default), provide column
percentages}

\item{pcTot}{\code{TRUE} or \code{FALSE} (default), provide total
percentages}

\item{barplot}{\code{TRUE} or \code{FALSE} (default), show barplots}

\item{yaxis}{ycounts (default) or ypc. Use respectively \code{counts} or
\code{percentages} for the bar plot y-axis}

\item{yaxisPc}{total_pc (default), column_pc, or row_pc. Use respectively
percentages \code{of total}, \code{within columns}, or \code{within rows}
for the bar plot y-axis.}

\item{xaxis}{rows (default), or columns in bar plot X axis}

\item{bartype}{stack or side by side (default), barplot type}

\item{formula}{(optional) the formula to use, see the examples}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$freqs} \tab \tab \tab \tab \tab a table of proportions \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$ctt} \tab \tab \tab \tab \tab a table \cr
  \code{results$cttma} \tab \tab \tab \tab \tab a table \cr
  \code{results$ctt2} \tab \tab \tab \tab \tab a table \cr
  \code{results$ctt3} \tab \tab \tab \tab \tab a table \cr
  \code{results$plotc} \tab \tab \tab \tab \tab an image \cr
  \code{results$barplot} \tab \tab \tab \tab \tab an image \cr
  \code{results$tabText} \tab \tab \tab \tab \tab a html \cr
  \code{results$SupportTab} \tab \tab \tab \tab \tab a table \cr
  \code{results$MoretabText} \tab \tab \tab \tab \tab a html \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$freqs$asDF}

\code{as.data.frame(results$freqs)}
}
\description{
Analysis of a 2 x 2 contingency table using relative risk, choose null and 
alternative values
}
\examples{
dat <- data.frame(
  Treatment = c('Folic acid', 'Folic acid', 'Placebo', 'Placebo'),
  Defect = c('Yes', 'No','Yes', 'No'),
  count = c(6, 587, 21, 581),
  check.names=FALSE)

  jeva::llr(formula = count ~ Treatment:Defect, data = dat, text = FALSE)
  #
  # RELATIVE RISK
  #
  # Contingency Tables
  # --------------------------------------
  #   Treatment     No      Yes    Total
  # --------------------------------------
  #   Folic acid     587      6      593
  #   Placebo        581     21      602
  #   Total         1168     27     1195
  # --------------------------------------
  #
  #
  # Support: Risk Ratio analyses
  # -----------------------------------------------------------------------------------------------------
  #                       RR Value    Difference     S             Param    G           df    p
  # -----------------------------------------------------------------------------------------------------
  #   H₀ vs risk ratio    1.000000    -0.02566082    -3.8944253     1, 2    8.788851     1    0.0030308
  #   Ha vs risk ratio    1.000000    -0.02566082    -4.3944253     2, 2    8.788851     1    0.0030308
  #   Ha vs H₀                         0.00000000    -0.5000000     2, 1    0.000000     1    1.0000000
  # -----------------------------------------------------------------------------------------------------
  #   Note. S uses Occam's Bonus correction for parameters (Param).
  #
  #
  # Support: Marginal main effects and interaction analyses, against the Null model
  # -------------------------------------------------------------------------------------------------------
  #   Component               Expected value    S              Param    G                df    p
  # -------------------------------------------------------------------------------------------------------
  #   Treatment                     597.5000     -0.4661085     2, 1       0.06778307     1     0.7945924
  #   Defect                        597.5000    698.7864811     2, 1    1398.57296214     1    < .0000001
  #   Treatment  ⨯  Defect                        3.8944260     2, 1       8.78885199     1     0.0030308
  #   Total                         298.7500    702.2147986     4, 1    1407.42959720     3    < .0000001
  # -------------------------------------------------------------------------------------------------------
  #   Note. S uses Occam's Bonus correction for parameters (Param).  The interaction and RR (against 1)
  #   will have the same S value. Adding the S values for the 3 components will precisely sum to the
  #   total S when no parameter correction is applied.
  #
  #
  # Intervals for RR
  # -------------------------------------------------------------------
  #   Type of interval      Level    RR          Lower       Upper
  # -------------------------------------------------------------------
  #   Support               2        1.025661    1.008768    1.037945
  #   Likelihood-based a    95\\%      1.025661    1.009130    1.037758
  # -------------------------------------------------------------------
  #   a See reference Pritikin et al (2017) such intervals are more
  #   accurate and are parameterization-invariant compared to
  #   conventional confidence intervals
  #

}
